pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hiveplugin"
version = "0.1.0"
authors = [
  { name = "Bhargav Garikipati", email = "bhargav.garikipati@stfc.ac.uk" },
]
description = "A plugin for working with Hive"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "idna==3.8",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "urllib3==2.2.2",
]

[project.scripts]
hiveplugin = "hiveplugin.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hiveplugin"]

[tool.hatch.metadata]
allow-direct-references = true